pyproject.toml
[tool.poetry]
name = "django-setting"
version = "0.1.0"
description = ""
authors = ["SaJH <wogur981208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
django-cors-headers = "^4.2.0"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.2.0"
pymysql = "^1.1.0"
pydantic = "^2.3.0"
pre-commit = "^3.4.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
